libhext: make use of clang's and GCC's -fvisibility parameter
When compiling with clang or GCC all symbols are now hidden by default.
Types and globals that are part of the API are explicitly set to public.
This reduces the filesize of libhext.so by 10% on my machine.

Two new macros:
* HEXT_PUBLIC: Every symbol that is supposed to be exposed to users is
  marked with HEXT_PUBLIC.
* HEXT_PRIVATE: All private member functions within public types are set
  to "hidden".
Defined in hext/Visibility.h.

When building libhext-test the visibility of libhext is overriden to
"default", i.e. everything is public. This is because libhext-test
contains unit tests accessing private symbols (e.g. TrimAndCollapseWs).