libhext: add 'greedy' option for rules
A rule marked with a plus sign does not stop at the first match, instead it
continually searches for a match until another rule takes precedence or until
there are no more elements left.
A greedy rule can also be made optional, that is, it may match none or many.