libhext: cmake: require boost min version 1.76.0
With boost v1.76.0 Boost.regex became a header-only library, removing the
need to link libboost_regex.{a,so}.
Unfortunately, some boost versions (like v1.81.0 on Debian 12)
incorrectly report that Boost.regex is unavailable, although the boost
header files (and therefore Boost.regex) are installed.
The easiest way to solve this is to require Boost with at least v1.76.0
and to remove any explicit dependencies on Boost.regex.